requirements.txt
Flask==1.1.2
mysql-connector-python==8.0.19
SQLAlchemy==1.3.16
PyMySQL==0.9.3

